Refactor logic to custom hook
